+++ apps/example-app/app/components/Icon.tsx
@@ -1,0 +1,94 @@
+import * as React from "react"
+import { ComponentType } from "react"
+import {
+  Image,
+  ImageStyle,
+  StyleProp,
+  TouchableOpacity,
+  TouchableOpacityProps,
+  View,
+  ViewProps,
+  ViewStyle,
+} from "react-native"
+
+export type IconTypes = keyof typeof iconRegistry
+
+interface IconProps extends TouchableOpacityProps {
+  /**
+   * The name of the icon
+   */
+  icon: IconTypes
+
+  /**
+   * An optional tint color for the icon
+   */
+  color?: string
+
+  /**
+   * An optional size for the icon. If not provided, the icon will be sized to the icon's resolution.
+   */
+  size?: number
+
+  /**
+   * Style overrides for the icon image
+   */
+  style?: StyleProp<ImageStyle>
+
+  /**
+   * Style overrides for the icon container
+   */
+  containerStyle?: StyleProp<ViewStyle>
+
+  /**
+   * An optional function to be called when the icon is pressed
+   */
+  onPress?: TouchableOpacityProps["onPress"]
+}
+
+/**
+ * A component to render a registered icon.
+ * It is wrapped in a <TouchableOpacity /> if `onPress` is provided, otherwise a <View />.
+ *
+ * - [Documentation and Examples](https://github.com/infinitered/ignite/blob/master/docs/Components-Icon.md)
+ */
+export function Icon(props: IconProps) {
+  const {
+    icon,
+    color,
+    size,
+    style: $imageStyleOverride,
+    containerStyle: $containerStyleOverride,
+    ...WrapperProps
+  } = props
+
+  const isPressable = !!WrapperProps.onPress
+  const Wrapper = (WrapperProps?.onPress ? TouchableOpacity : View) as ComponentType<
+    TouchableOpacityProps | ViewProps
+  >
+
+  const $imageStyle: StyleProp<ImageStyle> = [
+    $imageStyleBase,
+    color !== undefined && { tintColor: color },
+    size !== undefined && { width: size, height: size },
+    $imageStyleOverride,
+  ]
+
+  return (
+    <Wrapper
+      accessibilityRole={isPressable ? "imagebutton" : undefined}
+      {...WrapperProps}
+      style={$containerStyleOverride}
+    >
+      <Image style={$imageStyle} source={iconRegistry[icon]} />
+    </Wrapper>
+  )
+}
+
+export const iconRegistry = {
+  caretRight: require("../../assets/icons/caretRight.png"),
+  ladybug: require("../../assets/icons/ladybug.png"),
+}
+
+const $imageStyleBase: ImageStyle = {
+  resizeMode: "contain",
+}