+++ apps/example-app/app/components/index.ts
@@ -1,0 +1,6 @@
+export * from "./Button"
+export * from "./Icon"
+export * from "./ListItem"
+export * from "./ListView"
+export * from "./Screen"
+export * from "./Text"