+++ apps/example-app/babel.config.js
@@ -1,0 +1,17 @@
+/** @type {import('@babel/core').TransformOptions['plugins']} */
+const plugins = []
+
+/** @type {import('@babel/core').TransformOptions} */
+module.exports = function (api) {
+  api.cache(true)
+  return {
+    assumptions: {
+      setPublicClassFields: true,
+    },
+    presets: ["babel-preset-expo"],
+    env: {
+      production: {},
+    },
+    plugins,
+  }
+}