+++ apps/reactotron-app/src/renderer/contexts/Layout/useLayout.test.ts
@@ -1,4 +1,4 @@
-import { act, renderHook } from "@testing-library/react-hooks/dom"
+import { act, renderHook } from "@testing-library/react"
 
 import useLayout from "./useLayout"