+++ lib/reactotron-core-client/test/serialize.test.ts
@@ -16,6 +16,18 @@ test("serializes arrays", () => {
   expect(actual).toEqual(expected)
 })
 
+test("serializes BigInts", () => {
+  const mock = {
+    x: [BigInt(1), BigInt(2), BigInt(3)],
+    // eslint-disable-next-line @typescript-eslint/no-loss-of-precision, no-loss-of-precision
+    y: [BigInt(12345678901234567890), BigInt(2345434553442342345235243234)],
+  }
+  const actual = serialize(mock)
+  const expected = `{"x":["1","2","3"],"y":["12345678901234567168","2345434553442342324832043008"]}`
+
+  expect(actual).toEqual(expected)
+})
+
 test("serializes nested objects", () => {
   const mock = { x: { y: 1 } }
   const actual = serialize(mock)
@@ -49,6 +61,7 @@ test("medium sized funk", () => {
   }
   mockPayload.fn = hello
   mockPayload.anonymous = () => {}
+  mockPayload.bigInt = BigInt(1212)
 
   const actual = JSON.parse(serialize(mockPayload))
   expect(actual.string).toBe("String")
@@ -62,4 +75,5 @@ test("medium sized funk", () => {
   expect(actual.list[4]).toEqual({ foo: { x: 1 }, bar: "~~~ Circular Reference ~~~" })
   expect(actual.fn).toBe("~~~ hello() ~~~")
   expect(actual.anonymous).toBe("~~~ anonymous function ~~~")
+  expect(actual.bigInt).toBe("1212")
 })