+++ apps/reactotron-app/CHANGELOG.md
@@ -2,6 +2,66 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+### [3.7.2](https://github.com/infinitered/reactotron/compare/reactotron-app@3.7.1...reactotron-app@3.7.2) (2024-08-26)
+
+### [3.7.1](https://github.com/infinitered/reactotron/compare/reactotron-app@3.7.0...reactotron-app@3.7.1) (2024-04-04)
+
+## [3.7.0](https://github.com/infinitered/reactotron/compare/reactotron-app@3.6.2...reactotron-app@3.7.0) (2024-02-16)
+
+
+### Features
+
+* **reactotron-app:** Timeline deep-search ([#1417](https://github.com/infinitered/reactotron/issues/1417)) ([8d0e952](https://github.com/infinitered/reactotron/commit/8d0e9525ac30e895f38a0c2f99eeda155c14527f)), closes [#1181](https://github.com/infinitered/reactotron/issues/1181)
+
+### [3.6.2](https://github.com/infinitered/reactotron/compare/reactotron-app@3.6.1...reactotron-app@3.6.2) (2024-02-14)
+
+## [3.6.1](https://github.com/infinitered/reactotron/compare/reactotron-app@3.6.0...reactotron-app@3.6.1) (2024-02-14)
+
+
+### Bug Fixes
+
+* **reactotron-app:** added min width and height of electron window ([#1429](https://github.com/infinitered/reactotron/issues/1429)) ([a4467a2](https://github.com/infinitered/reactotron/commit/a4467a2d42a0532627d946578101379fe827436c))
+
+## [3.6.0](https://github.com/infinitered/reactotron/compare/reactotron-app@3.5.1...reactotron-app@3.6.0) (2024-02-14)
+
+
+### Features
+
+* **example-app:** Add a LOT of functionality to the demo app ([#1424](https://github.com/infinitered/reactotron/issues/1424)) [skip ci] ([41a539c](https://github.com/infinitered/reactotron/commit/41a539c7e56e968a83ca61b100468f5267865ddd))
+* **reactotron-app:** Added Electron BrowserWindow title for Linux users ([#1425](https://github.com/infinitered/reactotron/issues/1425)) ([726f255](https://github.com/infinitered/reactotron/commit/726f255412322598717f59f10d73ea7b4279f467))
+
+### [3.5.1](https://github.com/infinitered/reactotron/compare/reactotron-app@3.5.0...reactotron-app@3.5.1) (2024-01-25)
+
+## [3.5.0](https://github.com/infinitered/reactotron/compare/reactotron-app@3.4.0...reactotron-app@3.5.0) (2024-01-24)
+
+
+### Features
+
+* **reactotron-app:** Close the timeline search on 2nd “X” press ([#1418](https://github.com/infinitered/reactotron/issues/1418)) ([8e97074](https://github.com/infinitered/reactotron/commit/8e970743f50a54c6d3b411024e929b9a33dbcbe2)), closes [#1394](https://github.com/infinitered/reactotron/issues/1394)
+
+## [3.4.0](https://github.com/infinitered/reactotron/compare/reactotron-app@3.3.1...reactotron-app@3.4.0) (2024-01-24)
+
+
+### Features
+
+* **reactotron-app:** Adds tooltips to state snapshot item commands. ([#1386](https://github.com/infinitered/reactotron/issues/1386)) ([96891ac](https://github.com/infinitered/reactotron/commit/96891ac1ef94b63e0e5b292da3b5ddaa0e87e363))
+
+### [3.3.1](https://github.com/infinitered/reactotron/compare/reactotron-app@3.3.0...reactotron-app@3.3.1) (2024-01-23)
+
+## [3.3.0](https://github.com/infinitered/reactotron/compare/reactotron-app@3.2.0...reactotron-app@3.3.0) (2024-01-17)
+
+
+### Features
+
+* make home screen default route when app starts up ([#1393](https://github.com/infinitered/reactotron/issues/1393)) ([b78afb8](https://github.com/infinitered/reactotron/commit/b78afb89ba8e5e4805f07c7bc4dd6d9e3dd9e989))
+
+## [3.2.0](https://github.com/infinitered/reactotron/compare/reactotron-app@3.1.4...reactotron-app@3.2.0) (2024-01-15)
+
+
+### Features
+
+* **reactotron-app:** Move android devices from Help to Home ([#1396](https://github.com/infinitered/reactotron/issues/1396)) ([e7547c6](https://github.com/infinitered/reactotron/commit/e7547c6f36789e089bc2cfe1964e413268be2d05))
+
 ### [3.1.4](https://github.com/infinitered/reactotron/compare/reactotron-app@3.1.3...reactotron-app@3.1.4) (2023-12-22)